chore: Move "Upsert Track" function over
