+++ mobile/src/data/track/api.ts
@@ -3,6 +3,8 @@ import { eq } from "drizzle-orm";
 import { db } from "~/db";
 import { tracks } from "~/db/schema";
 
+import { getExcludedColumns } from "~/lib/drizzle";
+
 type InsertedTrack = typeof tracks.$inferInsert;
 
 //#region PATCH Methods
@@ -12,4 +14,33 @@ export async function updateTrack(
 ) {
   return db.update(tracks).set(values).where(eq(tracks.id, id));
 }
+//#endregion
+
+//#region PUT Methods
+/** Create/update track entries. */
+export function upsertTracks(entries: InsertedTrack[]) {
+  return db.insert(tracks).values(entries).onConflictDoUpdate({
+    target: tracks.id,
+    set: UpsertFields,
+  });
+}
+//#endregion
+
+//#region Internal Utils
+const UpsertFields = getExcludedColumns([
+  "name",
+  "rawArtistName", // ! This field is deprecated.
+  "albumId",
+  "disc",
+  "track",
+  "year",
+  "duration",
+  "format",
+  "bitrate",
+  "sampleRate",
+  "uri",
+  "modificationTime",
+  "fetchedArt",
+  "size",
+]);
 //#endregion
